+++ src/libs/conduit/conduit_node.cpp
@@ -15913,115 +15913,115 @@ Node::set_data_ptr(void *data)
     //release();
     m_data    = data;
 }
-
-namespace detail
-{
-    class AllocManager {
-     public:
-      static AllocManager& instance()
-      {
-        static AllocManager inst;
-        return inst;
-      }
-
-      AllocManager(AllocManager const&) = delete;
-      AllocManager(AllocManager&&) = delete;
-      AllocManager& operator=(AllocManager const&) = delete;
-      AllocManager& operator=(AllocManager &&) = delete;
-
-      index_t register_allocator(void*(*conduit_hnd_allocate) (size_t, size_t),
-                                 void(*conduit_hnd_free)(void *))
-      {
-          m_allocator_map[m_allocator_id] = conduit_hnd_allocate;
-          m_free_map[m_allocator_id]      = conduit_hnd_free;
-          std::cout << "I HEAR YOU " << (m_allocator_id+1) << std::endl;
-          return m_allocator_id++;
-
-      }
-
-      void * allocate(size_t n_items,
-                      size_t item_size,
-                      index_t allocator_id)
-      {
-        std::cout << "SURE THING USING " << allocator_id << std::endl;
-          return m_allocator_map[allocator_id](n_items, item_size);
-      }
-
-      void free(void *ptr,
-               index_t allocator_id)
-      {
-        std::cout << "LETS CLEAN THAT UP USING " << allocator_id << std::endl;
-          m_free_map[allocator_id](ptr);
-      }
-
-     protected:
-      AllocManager()
-          : m_allocator_map(),
-            m_free_map()
-      {
-          std::cout << "GIVE ME A SECOND ...." << std::endl;
-
-          m_allocator_map[0] = &conduit::utils::default_alloc_handler;
-          m_free_map[0]      = &conduit::utils::default_free_handler;
-          m_allocator_id = 1;
-          
-          std::cout << "I AM HERE TO HELP!" << std::endl;
-          // static std::map<index_t,void*(*)(size_t, size_t)> _allocator_map
-          //         = {{0, &default_alloc_handler}};
-          // return _allocator_map;
-          //
-          // static std::map<index_t,void(*)(void*)> _free_map
-          //     = {{0, default_free_handler}};
-          // return _free_map;
-      }
-
-      ~AllocManager()
-      {
-          std::cout << "SOMEONE HAS CALLED ME HOME!" << std::endl;
-      }
-
-      index_t m_allocator_id;
-      //-----------------------------------------------------------------------------
-      std::map<index_t,void*(*)(size_t, size_t)> m_allocator_map;
-      std::map<index_t,void(*)(void*)> m_free_map;
-
-     // And any other protected methods.
-    };
-}
-
-//-----------------------------------------------------------------------------
-index_t
-Node::register_allocator(void*(*conduit_hnd_allocate) (size_t, size_t),
-                         void(*conduit_hnd_free)(void *))
-{
-    return detail::AllocManager::instance().register_allocator(conduit_hnd_allocate,
-                                                               conduit_hnd_free);
-    // static index_t allocator_id = 1;
-    // allocator_map()[allocator_id] = conduit_hnd_allocate;
-    // free_map()[allocator_id]      = conduit_hnd_free;
-    // return allocator_id++;
-}
-
-
-//-----------------------------------------------------------------------------
-void *
-Node::conduit_allocate(size_t n_items,
-                       size_t item_size,
-                       index_t allocator_id)
-{
-  return detail::AllocManager::instance().allocate(n_items,item_size,
-                                              allocator_id);
-  // return allocator_map()[allocator_id](n_items, item_size);
-}
-
-//-----------------------------------------------------------------------------
-void
-Node::conduit_free(void *ptr,
-                   index_t allocator_id)
-{
-  detail::AllocManager::instance().free(ptr,allocator_id);
-  // free_map()[allocator_id](ptr);
-}
+//
+// namespace detail
+// {
+//     class AllocManager {
+//      public:
+//       static AllocManager& instance()
+//       {
+//         static AllocManager inst;
+//         return inst;
+//       }
+//
+//       AllocManager(AllocManager const&) = delete;
+//       AllocManager(AllocManager&&) = delete;
+//       AllocManager& operator=(AllocManager const&) = delete;
+//       AllocManager& operator=(AllocManager &&) = delete;
+//
+//       index_t register_allocator(void*(*conduit_hnd_allocate) (size_t, size_t),
+//                                  void(*conduit_hnd_free)(void *))
+//       {
+//           m_allocator_map[m_allocator_id] = conduit_hnd_allocate;
+//           m_free_map[m_allocator_id]      = conduit_hnd_free;
+//           std::cout << "I HEAR YOU " << (m_allocator_id+1) << std::endl;
+//           return m_allocator_id++;
+//
+//       }
+//
+//       void * allocate(size_t n_items,
+//                       size_t item_size,
+//                       index_t allocator_id)
+//       {
+//         std::cout << "SURE THING USING " << allocator_id << std::endl;
+//           return m_allocator_map[allocator_id](n_items, item_size);
+//       }
+//
+//       void free(void *ptr,
+//                index_t allocator_id)
+//       {
+//         std::cout << "LETS CLEAN THAT UP USING " << allocator_id << std::endl;
+//           m_free_map[allocator_id](ptr);
+//       }
+//
+//      protected:
+//       AllocManager()
+//           : m_allocator_map(),
+//             m_free_map()
+//       {
+//           std::cout << "GIVE ME A SECOND ...." << std::endl;
+//
+//           m_allocator_map[0] = &conduit::utils::default_alloc_handler;
+//           m_free_map[0]      = &conduit::utils::default_free_handler;
+//           m_allocator_id = 1;
+//
+//           std::cout << "I AM HERE TO HELP!" << std::endl;
+//           // static std::map<index_t,void*(*)(size_t, size_t)> _allocator_map
+//           //         = {{0, &default_alloc_handler}};
+//           // return _allocator_map;
+//           //
+//           // static std::map<index_t,void(*)(void*)> _free_map
+//           //     = {{0, default_free_handler}};
+//           // return _free_map;
+//       }
+//
+//       ~AllocManager()
+//       {
+//           std::cout << "SOMEONE HAS CALLED ME HOME!" << std::endl;
+//       }
+//
+//       index_t m_allocator_id;
+//       //-----------------------------------------------------------------------------
+//       std::map<index_t,void*(*)(size_t, size_t)> m_allocator_map;
+//       std::map<index_t,void(*)(void*)> m_free_map;
+//
+//      // And any other protected methods.
+//     };
+// }
+//
+// //-----------------------------------------------------------------------------
+// index_t
+// Node::register_allocator(void*(*conduit_hnd_allocate) (size_t, size_t),
+//                          void(*conduit_hnd_free)(void *))
+// {
+//     return detail::AllocManager::instance().register_allocator(conduit_hnd_allocate,
+//                                                                conduit_hnd_free);
+//     // static index_t allocator_id = 1;
+//     // allocator_map()[allocator_id] = conduit_hnd_allocate;
+//     // free_map()[allocator_id]      = conduit_hnd_free;
+//     // return allocator_id++;
+// }
+//
+//
+// //-----------------------------------------------------------------------------
+// void *
+// Node::conduit_allocate(size_t n_items,
+//                        size_t item_size,
+//                        index_t allocator_id)
+// {
+//   return detail::AllocManager::instance().allocate(n_items,item_size,
+//                                               allocator_id);
+//   // return allocator_map()[allocator_id](n_items, item_size);
+// }
+//
+// //-----------------------------------------------------------------------------
+// void
+// Node::conduit_free(void *ptr,
+//                    index_t allocator_id)
+// {
+//   detail::AllocManager::instance().free(ptr,allocator_id);
+//   // free_map()[allocator_id](ptr);
+// }
